:recycle: : moves stack feature to its own package
