+++ src/main/java/io/codeka/gaia/runner/StackRunner.java
@@ -5,8 +5,12 @@ import com.spotify.docker.client.exceptions.DockerException;
 import com.spotify.docker.client.messages.ContainerConfig;
 import io.codeka.gaia.bo.*;
 import io.codeka.gaia.modules.bo.TerraformModule;
-import io.codeka.gaia.repository.JobRepository;
-import io.codeka.gaia.repository.StackRepository;
+import io.codeka.gaia.stacks.repository.JobRepository;
+import io.codeka.gaia.stacks.repository.StackRepository;
+import io.codeka.gaia.stacks.bo.Job;
+import io.codeka.gaia.stacks.bo.JobType;
+import io.codeka.gaia.stacks.bo.Stack;
+import io.codeka.gaia.stacks.bo.StackState;
 import org.apache.commons.io.output.WriterOutputStream;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
