+++ src/main/java/io/codeka/gaia/stacks/bo/JobStatus.java
@@ -1,4 +1,4 @@
-package io.codeka.gaia.bo;
+package io.codeka.gaia.stacks.bo;
 
 public enum JobStatus {
     RUNNING, FINISHED, FAILED
