+++ src/main/java/io/codeka/gaia/stacks/bo/JobType.java
@@ -1,4 +1,4 @@
-package io.codeka.gaia.bo;
+package io.codeka.gaia.stacks.bo;
 
 public enum JobType {
     PREVIEW, RUN, STOP
