+++ src/main/java/io/codeka/gaia/stacks/repository/StackRepository.java
@@ -1,7 +1,7 @@
-package io.codeka.gaia.repository;
+package io.codeka.gaia.stacks.repository;
 
-import io.codeka.gaia.bo.Stack;
-import io.codeka.gaia.bo.StackState;
+import io.codeka.gaia.stacks.bo.Stack;
+import io.codeka.gaia.stacks.bo.StackState;
 import io.codeka.gaia.teams.bo.Team;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
