+++ src/test/java/io/codeka/gaia/runner/StackCommandBuilderTest.java
@@ -3,7 +3,7 @@ package io.codeka.gaia.runner;
 import com.github.mustachejava.DefaultMustacheFactory;
 import com.github.mustachejava.Mustache;
 import io.codeka.gaia.bo.Settings;
-import io.codeka.gaia.bo.Stack;
+import io.codeka.gaia.stacks.bo.Stack;
 import io.codeka.gaia.modules.bo.TerraformModule;
 import io.codeka.gaia.modules.bo.TerraformVariable;
 import org.jetbrains.annotations.NotNull;
