+++ src/test/java/io/codeka/gaia/runner/StackRunnerTest.java
@@ -7,8 +7,9 @@ import com.spotify.docker.client.messages.ContainerCreation;
 import com.spotify.docker.client.messages.ContainerExit;
 import io.codeka.gaia.bo.*;
 import io.codeka.gaia.modules.bo.TerraformModule;
-import io.codeka.gaia.repository.JobRepository;
-import io.codeka.gaia.repository.StackRepository;
+import io.codeka.gaia.stacks.repository.JobRepository;
+import io.codeka.gaia.stacks.repository.StackRepository;
+import io.codeka.gaia.stacks.bo.*;
 import io.codeka.gaia.teams.bo.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
