+++ src/test/java/io/codeka/gaia/stacks/repository/JobRepositoryIT.java
@@ -1,7 +1,8 @@
-package io.codeka.gaia.repository;
+package io.codeka.gaia.stacks.repository;
 
-import io.codeka.gaia.bo.Job;
-import io.codeka.gaia.bo.JobStatus;
+import io.codeka.gaia.stacks.bo.Job;
+import io.codeka.gaia.stacks.bo.JobStatus;
+import io.codeka.gaia.stacks.repository.JobRepository;
 import io.codeka.gaia.teams.bo.User;
 import io.codeka.gaia.teams.repository.UserRepository;
 import io.codeka.gaia.test.MongoContainer;
